Working payload encryption and integrity verification (#3)
# Added features / functionality:
- Proper encryption / decryption 
- Payload integrity verification
    - At predefined intervals
    - On method change
- TCP_ACK_Bounce now works too
 
## Encryption / Decryption

Payload is encrypted and subsequently decrypted correctly, stripping the AES padding from the end of the message too

## Payload Integrity

Managing payload integrity using a `beaconing` process, a sort of challenge and response that does not require the receiver (beacon sender) to know the IP address of the sender - knowing the sender IP is not always a given, for example when using the TCP Ack bounce technique.

Using an ARP packet with the payload (A single verification byte) as the host address (y.y.y.x) where y matches the network of the receiver and x is the payload. This technique is almost guaranteed to be successful since blocking / altering the ARP packet would be fatal to general communications.

The challenge is done either when the method changes (to confirm the previous methods data was received correctly) or every x packets.

## TCP ACK Bounce implementation

TCP_ACK_Bounce is now working, previously not due to attempting to use the receivers MAC address, using the senders MAC works.

In the test environment I am using TCP Replay to simulate traffic, so the TCP servers detected by the tool are not valid (but would be in an actual network), instead I have set up a dummy interface that will perform the functions required, this is hardcoded into the response of `get_tcp_server`.